Build example projects (#76)
Build the Examples in CI.

### Motivation:

`swift-metrics-extras` now has an `Example/ServiceIntegration`, which
depends on the package, but is not verified in the CI.

### Modifications:

- `Scripts/test-examples.sh` added (inspired by the
[`swift-configuration/Scripts/test-examples.sh`](https://github.com/apple/swift-configuration/blob/main/Scripts/test-examples.sh),
but the structure is different and this one is more simple).
- `pull_request` and `main` workflows use shared
`apple/swift-nio/.github/workflows/swift_test_matrix.yml@main` to run
the script.

### Result:

`Example/ServiceIntegration` is being built in CI on Linux