Add a Grafana dashboard to the Service Integration example (#77)
Add Grafana integration with a provisioned Process System Metrics
dashboard to the Service Integration example.

### Motivation:

Process System Metrics collected and exported in this package ultimately
should be used in a monitoring dashboard. While it is common for
operates to be opinionated about the their dashboards, we also want to
help new users with observability adoption by providing an example
dashboard they can use as an inspiration or directly add to their
project.

### Modifications:

- OTel exporting is added to the `ServiceIntegration` example via
`swift-otel` package.
- A simple Grafana dashboard utilizing the metrics implemented by the
`SystemMetricsMonitor` added to the example.
- A docker-compose deployment added to the example.

### Result:

Users can deploy a sample Grafana dashboard with SystemMetricsMonitor
metrics and use it as an inspiration for creating monitoring for their
services.

<img width="1118" height="1042" alt="Screenshot 2025-12-17 at 17 38 23"
src="https://github.com/user-attachments/assets/584735c0-8823-4b5b-a609-1b8856ee3cc3"
/>